+++ Editor/Scripts/kebinImports.cs
@@ -205,13 +205,13 @@ namespace kebinImports
             EditorApplication.update -= UpdateSettings;
             hideWarnings = EditorPrefs.GetBool("kebinImports.hideWarnings", false);
             showSplash = EditorPrefs.GetBool("kebinImports.showSplash", true);
-            avatarEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.avatarEssentials", "1000000000"));
+            avatarEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.avatarEssentials", "10000000000"));
             worldEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.worldEssentials", "1000"));
-            if (avatarEssentials.Length != 7 && worldEssentials.Length != 4)
+            if (avatarEssentials.Length != AENames.Length && worldEssentials.Length != WENames.Length)
             {
-                EditorPrefs.SetString("kebinImports.avatarEssentials", "1000000000");
+                EditorPrefs.SetString("kebinImports.avatarEssentials", "10000000000");
                 EditorPrefs.SetString("kebinImports.worldEssentials", "1000");
-                avatarEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.avatarEssentials", "1000000000"));
+                avatarEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.avatarEssentials", "10000000000"));
                 worldEssentials = Str2BoolArr(EditorPrefs.GetString("kebinImports.worldEssentials", "1000"));
             }
             if (Directory.Exists(Application.dataPath + "/../Packages/com.vrchat.base"))